Safari-611.1.10/LayoutTests/fast/canvas/webgl/bufferData-offset-length-expected.txt

 Checks the last two arguments to bufferData() are respected.

PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS Left half of canvas should be filled
PASS Right half of canvas should be empty
PASS gl.getError() is gl.NO_ERROR
PASS Left half of canvas should be empty
PASS Right half of canvas should be filled
PASS gl.getError() is gl.NO_ERROR
PASS Left half of canvas should be empty
PASS Right half of canvas should be filled
PASS gl.getError() is gl.NO_ERROR
PASS Left half of canvas should be filled
PASS Right half of canvas should be empty
PASS successfullyParsed is true

TEST COMPLETE
